fixing possible NULL dereferencing as found by static code analyzer Remove many declarations of type long to avoid 64-bit problems Make casting from 64 bit to 32 bit explicit
git-svn-id: svn://svn.code.sf.net/p/rtf2latex2e/code@39 ec8729b6-1b83-4082-b6c8-10560a4245e1

+++ src/rtf.h
@@ -323,7 +323,7 @@ typedef	void (*RTFFuncPtr) ();		/* generic function pointer */
  */
 
 void		RTFSimpleInit ();
-void 		RTFSetDefaultFont (long fontNumber);
+void 		RTFSetDefaultFont (int fontNumber);
 void		RTFInit ();
 void 		RTFSetStream(FILE *stream);
 char 		*RTFGetInputName();
